Initial version of separation of remote and embedded API

+++ core/src/main/java/com/jetbrains/youtrack/db/internal/core/sql/filter/SQLFilterCondition.java
@@ -26,13 +26,13 @@ import com.jetbrains.youtrack.db.api.query.Result;
 import com.jetbrains.youtrack.db.api.record.Identifiable;
 import com.jetbrains.youtrack.db.api.record.RID;
 import com.jetbrains.youtrack.db.api.schema.Collate;
-import com.jetbrains.youtrack.db.internal.core.metadata.schema.PropertyTypeInternal;
 import com.jetbrains.youtrack.db.internal.common.collection.MultiValue;
 import com.jetbrains.youtrack.db.internal.common.log.LogManager;
 import com.jetbrains.youtrack.db.internal.core.command.CommandContext;
 import com.jetbrains.youtrack.db.internal.core.db.DatabaseSessionInternal;
 import com.jetbrains.youtrack.db.internal.core.exception.QueryParsingException;
 import com.jetbrains.youtrack.db.internal.core.id.RecordId;
+import com.jetbrains.youtrack.db.internal.core.metadata.schema.PropertyTypeInternal;
 import com.jetbrains.youtrack.db.internal.core.query.QueryRuntimeValueMulti;
 import com.jetbrains.youtrack.db.internal.core.record.impl.EntityImpl;
 import com.jetbrains.youtrack.db.internal.core.serialization.serializer.StringSerializerHelper;
@@ -477,7 +477,7 @@ public class SQLFilterCondition {
       final Iterable<?> multiValue = MultiValue.getMultiValueIterable(iValue);
 
       // MULTI VALUE: RETURN A COPY
-      final var result = new ArrayList<>(MultiValue.getSize(iValue));
+      final var result = new ArrayList<>((int) MultiValue.getSize(iValue));
 
       for (final var value : multiValue) {
         if (value instanceof SQLFilterItem) {
