+++ core/src/main/java/com/jetbrains/youtrack/db/internal/core/sql/functions/misc/SQLMethodInclude.java
@@ -112,7 +112,7 @@ public class SQLMethodInclude extends AbstractSQLMethod {
         return copy((Map) iThis, iParams, session);
       } else if (MultiValue.isMultiValue(iThis)) {
         // ACT ON MULTIPLE DOCUMENTS
-        final List<Object> result = new ArrayList<Object>(MultiValue.getSize(iThis));
+        final List<Object> result = new ArrayList<Object>((int) MultiValue.getSize(iThis));
         for (var o : MultiValue.getMultiValueIterable(iThis)) {
           if (o instanceof Identifiable) {
             try {
